Moved FlushEventListener and FlushMetrics to private API
Moved the auto-flush specific classes - FlushEventListener and FlushMetrics to private API.
Also, removed the AutoFlushController from StreamLayerClientImpl

Relates to: OLPEDGE-826
