+++ olp-cpp-sdk-dataservice-write/src/FlushEventListener.h
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include <olp/dataservice/write/FlushMetrics.h>
+#include "FlushMetrics.h"
 
 namespace olp {
 namespace dataservice {
