+++ tests/common/CMakeLists.txt
@@ -21,7 +21,6 @@ set(OLP_SDK_TESTS_COMMON_HEADERS
     ${CMAKE_CURRENT_SOURCE_DIR}/matchers/NetworkUrlMatchers.h
     ${CMAKE_CURRENT_SOURCE_DIR}/mocks/NetworkMock.h
     ${CMAKE_CURRENT_SOURCE_DIR}/mocks/CacheMock.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/testables/FlushEventListenerTestable.h
 )
 
 set(OLP_SDK_TESTS_COMMON_SOURCES
